requirements.txt
ffmpeg-python
picamera
RPi.GPIO
abc
utils